pyproject.toml
[project]
name = "drive-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-auth",
    "google-auth-oauthlib", 
    "google-auth-httplib2",
    "google-api-python-client",
]
